Rails 5.1.5 update

+++ app/controllers/user_sessions_controller.rb
@@ -5,7 +5,8 @@ class UserSessionsController < ApplicationController
   end
 
   def create
-    @user_session = UserSession.new(user_session_params)
+    
+    @user_session = UserSession.new(user_session_params.to_h)
     if @user_session.save
       flash[:notice] = "Hallo #{current_user.login}, schön dass Du da bist."
       redirect_to root_url
@@ -16,13 +17,13 @@ class UserSessionsController < ApplicationController
   end
 
   def destroy
-    current_user_session.destroy
+    current_user_session.destroy unless current_user_session.nil?
     redirect_to new_user_session_url
   end
 
   private
 
     def user_session_params
-      params.require(:user_session).permit(:login, :password)
+      params.require(:user_session).permit(:login, :password, :remember_me)
     end
 end
